console: Reduce size of fields in PrintConsole struct
This reduces the size of the struct from 88 to 68 bytes.

It isn't feasible to reorganize the fields to reduce padding because
there may be people using this struct and not assigning the fields by
name but by order.

    PrintConsole console = {
        0,
        NULL,
        1
     };

Vs.

    PrintConsole console = {
        .field1 = 0,
        .pointer = NULL,
        .field2 = 1
     };

+++ include/nds/arm9/console.h
@@ -126,21 +126,21 @@ typedef struct PrintConsole
     u8 bgLayer;         ///< Bg layer used by the background
     int bgId;           ///< bgId, should be set with a call to bgInit() or bgInitSub()
 
-    int cursorX;        ///< Current X location of the cursor (as a tile offset by default)
-    int cursorY;        ///< Current Y location of the cursor (as a tile offset by default)
+    s16 cursorX;        ///< Current X location of the cursor (as a tile offset by default)
+    s16 cursorY;        ///< Current Y location of the cursor (as a tile offset by default)
 
-    int prevCursorX;    ///< Internal state
-    int prevCursorY;    ///< Internal state
+    s16 prevCursorX;    ///< Internal state
+    s16 prevCursorY;    ///< Internal state
 
-    int consoleWidth;   ///< Width of the console hardware layer in tiles
-    int consoleHeight;  ///< Height of the console hardware layer in tiles
+    u16 consoleWidth;   ///< Width of the console hardware layer in tiles
+    u16 consoleHeight;  ///< Height of the console hardware layer in tiles
 
-    int windowX;        ///< Window X location in tiles
-    int windowY;        ///< Window Y location in tiles
-    int windowWidth;    ///< Window width in tiles
-    int windowHeight;   ///< Window height in tiles
+    u16 windowX;        ///< Window X location in tiles
+    u16 windowY;        ///< Window Y location in tiles
+    u16 windowWidth;    ///< Window width in tiles
+    u16 windowHeight;   ///< Window height in tiles
 
-    int tabSize;        ///< Size of a tab
+    u8 tabSize;        ///< Size of a tab
 
     u16 fontCharOffset; ///< Offset to the first graphics tile in background
                         /// memory (in case your font is not loaded at a
